Moved main to cmd/influx-spout
The main package for influx-spout is now at cmd/influx-spout as per
typical Go source code arragement idioms. The end-to-end test has been
moved there too which makes more sense than it being in the spouttest
package.
